PR1

+++ src/user/user.service.ts
@@ -1,14 +1,21 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectRepository } from '@nestjs/typeorm';
 import { User } from './entities/user.entity';
 import { Repository } from 'typeorm';
+import * as bcrypt from 'bcrypt';
+import { ConfigService } from '@nestjs/config';
 
 @Injectable()
 export class UserService {
   constructor(
     @InjectRepository(User) private readonly userRepository: Repository<User>,
+    private readonly configService: ConfigService,
   ) {}
 
   async validateExists(id: number) {
@@ -19,8 +26,27 @@ export class UserService {
     }
   }
 
-  create(createUserDto: CreateUserDto) {
-    return this.userRepository.save(createUserDto);
+  async create(createUserDto: CreateUserDto) {
+    const { email, password } = createUserDto;
+
+    const user = await this.userRepository.findOne({
+      where: { email },
+    });
+
+    if (user) {
+      throw new BadRequestException('이미 가입한 이메일 입니다.');
+    }
+
+    const hashRounds = this.configService.get<number>('HASH_ROUNDS');
+    if (!hashRounds) {
+      throw new Error('HASH_ROUNDS 환경 변수가 설정되지 않았습니다.');
+    }
+
+    const hashedPassword = await bcrypt.hash(password, hashRounds);
+
+    await this.userRepository.save({ email, password: hashedPassword });
+
+    return this.userRepository.findOne({ where: { email } });
   }
 
   async findAll() {
@@ -51,8 +77,14 @@ export class UserService {
   }
 
   async remove(id: number) {
-    await this.validateExists(id);
+    const user = await this.userRepository.findOne({ where: { id } });
 
-    return await this.userRepository.delete(id);
+    if (!user) {
+      throw new NotFoundException(`아이디가 ${id}인 유저가 없습니다.`);
+    }
+
+    await this.userRepository.delete(id);
+
+    return id;
   }
 }
